+++ nengo_dl/tests/test_utils.py
@@ -1,7 +1,6 @@
 # pylint: disable=missing-docstring
 
 import numpy as np
-from packaging import version
 import tensorflow as tf
 
 from nengo_dl import utils
@@ -67,9 +66,6 @@ def test_progress_bar():
 
 
 def test_gpu_check():
-    if version.parse(tf.__version__) < version.parse("2.1.0rc0"):
-        gpus_available = tf.config.experimental.list_physical_devices("GPU")
-    else:
-        gpus_available = tf.config.list_physical_devices("GPU")
+    gpus_available = tf.config.list_physical_devices("GPU")
 
     assert utils.tf_gpu_installed == (len(gpus_available) > 0)