+++ nengo_dl/utils.py
@@ -11,7 +11,6 @@ import time
 
 from nengo.exceptions import SimulationError
 import numpy as np
-from packaging import version
 import progressbar
 import tensorflow as tf
 
@@ -28,12 +27,7 @@ tf_gpu_installed = not subprocess.call(
         "-c",
         "import sys; "
         "import tensorflow as tf; "
-        "sys.exit(len(tf.config%s.list_physical_devices('GPU')) == 0)"
-        % (
-            ".experimental"
-            if version.parse(tf.__version__) < version.parse("2.1.0rc0")
-            else ""
-        ),
+        "sys.exit(len(tf.config.list_physical_devices('GPU')) == 0)",
     ]
 )